model	R Documentation

Data generation template and analysis template for simulation.

Description

Creates a data generation and analysis template (lavaan parameter table) for simulations with structural equation models based on Y-side LISREL design matrices. Each corresponds to a LISREL matrix, but must be a SimMatrix or SimVector built using bind. In addition to the usual Y-side matrices in LISREL, both PS and TE can be specified using correlations (RPS, RTE) and scaled by a vector of residual variances (VTE, VPS) or total variances (VY, VE). Multiple groups are supported by passing lists of SimMatrix or SimVector to arguments, or by specifying the number of groups.

Usage

model(LY = NULL, PS = NULL, RPS = NULL, TE = NULL, RTE = NULL, BE = NULL, 
VTE = NULL, VY = NULL, VPS = NULL, VE = NULL, TY = NULL, AL = NULL, MY = NULL, 
ME = NULL, modelType, indLab=NULL, facLab=NULL, ngroups=1, smartStart=TRUE)

Arguments

LY
Factor loading matrix from endogenous factors to Y indicators (need to be SimMatrix object).

PS
Residual covariance matrix among endogenous factors (need to be SimMatrix object).

RPS
Residual correlation matrix among endogenous factors (need to be SimMatrix object).

TE
Measurement error covariance matrix among Y indicators (need to be SimMatrix object).

RTE
Measurement error correlation matrix among Y indicators (need to be SimMatrix object).

BE
Regression coefficient matrix among endogenous factors (need to be SimMatrix object).

VTE
Measurement error variance of indicators (need to be SimVector object).

VY
Total variance of indicators (need to be SimVector object). NOTE: Either measurement error variance or indicator variance is specified. Both cannot be simultaneously specified.

VPS
Residual variance of factors (need to be SimVector object).

VE
Total variance of of factors (need to be SimVector object). NOTE: Either residual variance of factors or total variance of factors is specified. Both cannot be simulatneously specified.

TY
Measurement intercepts of Y indicators. (need to be SimVector object).

AL
Endogenous factor intercept (need to be SimVector object).

MY
Overall Y indicator means. (need to be SimVector object). NOTE: Either measurement intercept of indicator mean can be specified. Both cannot be specified simultaneously.

ME
Total mean of endogenous factors (need to be SimVector object). NOTE: Either endogenous factor intercept or total mean of endogenous factor is specified. Both cannot be simultaneously specified.

modelType
"CFA", "Sem", or "Path". This is specified to ensure that the analysis and data generation template created based on specified matrices in model correspond to what the user intends.

indLab
Character vector of indicator labels. If left blank, automatic labels will be generated as x1, x2, ... xx.

facLab
Character vector of factor labels. If left blank, automatic labels will be generated as y1, y2, ... yx

ngroups
Integer. Number of groups for data generation, defaults to 1. If larger than one, all specified matrices will be repeated for each additional group. If any matrix argument is a list, the length of this list will be the number of groups and ngroups is ignored.

smartStart
Defaults to FALSE. If TRUE, population parameter values that are real numbers will be used as starting values.

Details

The simsem package is intricately tied to the lavaan package for analysis of structural equation models. The analysis template that is generated by model is a lavaan parameter table, a low-level access point to lavaan that allows repeated analyses to happen more rapidly. If desired, the parameter table generated can be used directly with lavaan for many analyses.

The data generation template is simply a list of SimMatrix or SimVector. The SimSem object can be passed to the function generate to generate data.

If multiple group data is desired, the user can optionally either specify the number of groups argument, or pass a list of SimMatrix or SimVector to any of the matrix arguments. The length of this list will be the number of groups. If only one argument is a list, all other arguments will be automatically replicated to that length, parameters will be identified in the same way, have the same population parameter value/distribution, and have the same misspecification. If only ngroups is specified, all arguments will be replicated in this fashion. If equality constraints are present during the automatic replication, these parameters will be constrained to be equal across groups.

Value

SimSem object that contains the data generation template (@dgen) and analysis template (@pt).

Author(s)

Patrick Miller (Univeristy of Notre Dame; pmille13@nd.edu), Sunthud Pornprasertmanit (University of Kansas; psunthud@ku.edu)

See Also

sim for simulations using the SimSem template.

generate To generate data using the SimSem template.

analyze To analyze real or generated data using the SimSem template.

draw To draw parameters using the SimSem template.

Examples

loading <- matrix(0, 6, 2)
loading[1:3, 1] <- NA
loading[4:6, 2] <- NA
LY <- bind(loading, 0.7)

latent.cor <- matrix(NA, 2, 2)
diag(latent.cor) <- 1
RPS <- binds(latent.cor, 0.5)

RTE <- binds(diag(6))

VY <- bind(rep(NA,6),2)

CFA.Model <- model(LY = LY, RPS = RPS, RTE = RTE, modelType = "CFA")